requirements.txt
aiogram>=3.2.0
python-dotenv>=1.0.0
aiohttp>=3.9.1
redis>=5.0.1  # Optional: for FSM storage 